Allow `flex-shrink` to be set to 0.
The `flex-shrink` css property could not be set to 0 since Mesop will
ignore styles that are "empty".

Normally this is fine, but `flex-shrink` defaults to 1, which means the
"empty" state will make `flex-shrink` equal to 1. This means that 0
cannot be set for `flex-shrink`.

To resolve this, we can change the date type to a string in the UI
proto.

Example:

```
import mesop as me

@me.page(path="/flex_shrink")
def main():
  with me.box(style=me.Style(display="flex", flex_shrink="0")):
    me.text("Testing flex shrink")
```