Add component diffing on event responses.
This change is gated by a flag called `enable_component_tree_diffs`.
When enabled, the full component tree will not be returned in event
responses. Instead only the changed compnent fields will be returned
to the frontend. The frontend will then apply the diff to update its
copy of the component tree.