Add Uploader Component
This component will render a basic file input. On upload the component
will send the data to the Mesop server where the user can then validate
the data and use it for their application.

Use cases include multi-modal use cases where you want to upload an
image, audio, or video to an multimodal model.

The uploader does not support the following features:

- Drag and drop
- Image/audio/video preview
- Multi file upload
- Does not try to limit file size on the frontend
- Image will be uploaded immediately on upload (i.e. no click event to
trigger the upload)
- Also no styling exposed yet.