Enable markdown tables + attribute lists
This change allows markdown tables to be rendered correctly. One caveat
is that custom styling is not possible unless someone includes an custom
stylesheet via CDN or external URL.

This means that the default rendering is very basic.

If a user wants to use render markdown from a pandas data frame, they
will need to install tabulate separately.

Attribute lists have also been enabled to make it easier to style the
the tables via CSS.