Add file backend for state sessions
This change also adds a new dependency with msgpack. For now we use it
to serialize/deserialize the JSON strings of the Mesop state. In the
future we will want to serialize everything with msgpack.

The reason we can't do this now is because we will need to add
msgpack encoding/decoding rules to support non-standard data types.

This change also adds a new config variable called
MESOP_STATE_SESSION_BACKEND_FILE_BASE_DIR which allows the user to
specify the location where the state sessions get written.