Add SQL state session backend.
This state session backend uses SQL Alchemy to support different SQL
databases.

I tested using SQlite3 and PostgreSQL. Both seemed to work fine.

One slight workflow issue is that for non-SQLite3 databases, people
will need to manually run a script to create the database table used
for storing state sessions. The commands are mentioned in the config.md
file. That's not the most user friendly approach. So probably need a
better way to approach it. But I think for the initial version of this
backend, it should be OK, especially considering we don't know how many
people will actually use it.