Fix bug with .env not loading for pip installs
For some reason the default `load_dotenv()` behavior is not working as
expected for pip installs. To fix this, we update the code to explicitly
search the user's current working directory.

This does not work when running in Bazel, so we still need to use
`load_dotenv()`.

Tested this by running ./scripts/pip.sh and also testing
./scripts/cli.sh.