Make UploadedFile class serializable (#679)
* Make UploadedFile class serializable

Since we're inheriting from io.Bytes, this made it harder to serialize
since we couldn't just make it a data class.

So to workaround this we added custom encoder/decoder logic to handle
UploadedFile specifically.

We also needed to add a custom operator for DeepDiff. For whatever
reason deep diff wasn't diffing it correctly, so instead we just perform
a basical equality check here.

Demos/tests have been updated to illustrate that UploadedFile is now
serializable.