Update some usages of scroll-to
We can remove the intentional delay on scroll-to now.

There is one case where this does not work as expected though. When the
chat message is not streamed, the scroll to with hidden box strategy
does not work since the output is too fast. This occurred on LLM
Rewriter example. But I updated it to fake the streaming.